refactor: split `elfs` crate
