+++ crates/proposer-vkeys-raw/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "proposer-vkeys-raw"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+op-succinct-elfs.workspace = true
+
+[dev-dependencies]
+alloy-primitives.workspace = true
+
+[build-dependencies]
+prover-elf-utils.workspace = true
+op-succinct-elfs.workspace = true
